Remove cast to `caddr_t` that breaks OS X/GCC build.
Currently the build is broken on OS X/GCC because we are passing an
integral type casted to `caddr_t` as the first argument to a call to
`mmap`.

`caddr_t` is used in some kernels (cf. BSD and Linux) to represent
byte-aligned virtual memory addresses. `caddr_t` was never adopted by
the POSIX specification, and in many OSs is not surfaced to the user.
Furthermore, the POSIX spec defines the `mmap` prototype to take `void*`
as the type of the first argument.

In this case the intent of the call is to pass `NULL` to `mmap`, which
would case the kernel to choose address at which to create the mapping.
This commit will therefore make the first argument of this call
`nullptr`, which is more portable.