Delete trailing whitespace
Since no-trailing-whitespace is in our style guide, my emacs is set
to wipe out trailing whitespace. But we have enough of it lying
around that this causes random lines to change in a lot of commits.

I'm wiping this out once and for all here. Well, at least until we
copy over junk from elsewhere.