Fix segfault due to mismatched DocumentFrequencyTable
This change is required for
a2e8d00951a0dbc6de1dcebcb5b59027a0b02171 to not cause LogAssert
failures.

This also fixes the test and does rounding to convert frequencies
to counts. The rounding may not be the correct way to round, but
it should be fine for the counts/frequencies we have in the test
at the moment.