Revert "Revert "Fix slice leak in RowTableIntegration test.""
This reverts commit 761b71b8945e7700b182d9837fa9bdbb4902e80c.

I believe the underlying issue that caused aa28796 to fail non-determistically
has been fixed. However, I was never able to reproduce the failure, so it's
possible that I'm wrong and that I'm re-introducing a subtle non-deterministic
bug. In the event that a non-determistic failure pops up in something that's
related to DocumentHandleTest or TrackingSliceBufferAllocator, we should be
highly suspicious of this commit.