Ignore -Wold-style-cast warnings for `MAP_FATAL` (#273)
In #233 we describe a build break affecting users of the default
toolchains of various popular Unix derivatives (notably Ubuntu 16.04,
whose default g++ package is version 5.04, and OS X 10.10.2, whose
default Clang distribution is Apple LLVM 6.1.0, clang-602.0.53).

The build break concerns a symbol `MAP_FATAL`, which is implemented in
these versions of libc using an old C-style cast (typically
`(void*)-1`). Since the BitFunnel build system treats C-style casts as
compile errors, simply referencing this symbol breaks the build on these
platforms.

In this commit, we will suppress the `-Wold-style-cast` warnings on the
specific liens we're referencing `MAP_FATAL`. Since these lines occur
only on the Unix builds, the Windows codepath is unaffected. The
specific implementation of this warning suppression involves a `#pragma
GCC`, which both Clang and GCC respect, so this commit will work for
both OS X and for Ubuntu.