Fix #225; work-around for build-breaking bug in some versions of GCC (#274)
In #225 we investigate and discuss in detail a build break that affects
some versions of GCC (5.4.0 and 5.4.1, to name a few). In brief, these
versions of GCC emit a warning that claims some variables might be
uninitialized, though the warnings seem to be spurious.

After some discussion we concluded that it is likely a bug with the `<<`
operator, tracked in GCC as bug #66977. Since this toolchain is the
default distribution of the `g++` package in apt for Ubuntu 16.04, so it
is of interest for us to work-around this bug.

This commit, therefore, suppresses this specific warning for the
specific lines that cause it.