Replace time-intensive correlation algorithm with space-intenstive algorithm.
Also adds TermToText lookup.

The previous algorithm looked up a set of rows for each pair of terms and
determined the size of the intersection. My original plan was to replace the
slow set data structure with something faster, but that wouldn't work because of
the slow speed of Term creation
(https://github.com/BitFunnel/BitFunnel/issues/311), and Term creation is
required to use the TermTable to get a RowIdSequence per Term. If you ignore the
cost of set intersection, this algorithm is O(n**2).

This algorithm creates a map of RowId->Term, and uses that to construct a map of
Term->[Term->collision count]. This algorithm is O(n) to create the map, but the
second map is O(n**2) space and walking through it to create the output is
O(n**2) run time. However, in practice, this version is much faster than the
previous version when the space isn't so large that we start paging to disk, and
in that case the previous algorithm is prohibitively slow. I believe this is
because while the algorithm is worst case O(n**2), in practice we have few
enough collisions that it's not actually O(n**2).

It's possible that the previous algorithm could be sped up by both speeding up
set intersection and caching terms so that we don't create Terms O(n**2) times.