Fixed dedupe buffer overflow in NativeCodeGenerator.
This change seems to eliminate the crash.
It runs 400k DFT queries from the first 15k documents of Wikipedia without crashing
and it passes all unit tests. We will need to give it more scrutiny before merging
into master. At the very least, we should run the verifier on 400k terms and perhaps
examine a few of the query log metrics.