Avoid dropping all but head in querylog, part 1.
See https://github.com/BitFunnel/BitFunnel/issues/342 for more discussion. We
have two mechanisms that are dropping everything but the head. The first is that
we're using a log-normal distribution with some specific constants and the
second is that we're dropping all but the top quarter of the terms. This avoids
dropping all but the top quarter but doesn't change the constants for the
log-normal distribution.