Synthetic index verification.
Added "verify synthetic" command to REPL. Command generates synthetic
queries and then verifies matches for these queries. Assumes an index
where a document with id X contains terms of the form "pY" where Y
is a prime factor of X. So, for example, document 99 would contain the
terms "p3" and "p11" since its prime factorization is 3*3*11.

The synthetic index is based on prime factorizations because
  1. There exists a simple way of determining which documents
     contain which terms.
  2. Prime factors have widely varying frequencies in the corpus.
     Contrast this with positions of set bits in the document id.
     These bits all have frequencies close to 0.5.